requirements.txt
fastapi==0.98.0
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
requests==2.31.0
